+++ src/utils/index.ts
@@ -68,3 +68,7 @@ export * from './Code/parse-highlight-lines.util'
 
 export * from './Textarea/sanitize-html.util'
 export * from './Textarea/html-to-markdown.util'
+
+export * from './Mask/apply-mask.util'
+export * from './Mask/resolve-mask-config.util'
+export * from './Mask/validate-pattern.util'